Created PlayerHealth class derived from Health.
